chore(airflow): merge earthquake classes
